requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
itsdangerous==2.2.0
PyYAML==6.0.2
